Refactor and clean-up CoreIR backend
