+++ magma/simulator/coreir_simulator.py
@@ -2,7 +2,7 @@ import os
 from tempfile import NamedTemporaryFile
 
 from .simulator import CircuitSimulator, ExecutionState
-from ..backend import coreir_, coreir_utils
+from magma.backend.coreir import coreir_backend, coreir_utils
 from ..frontend.coreir_ import GetMagmaContext
 from ..scope import Scope
 from ..ref import DefnRef, ArrayRef, TupleRef
@@ -154,7 +154,7 @@ class CoreIRSimulator(CircuitSimulator):
             self.ctx = GetMagmaContext()
         else:
             self.ctx = context
-        coreir_.compile(circuit, coreir_filename, context=self.ctx)
+        coreir_backend.compile(circuit, coreir_filename, context=self.ctx)
 
         # Initialize interpreter, get handle back to interpreter state
         self.ctx.get_lib("commonlib")
